fix(ci): python-coverage-shards — keep first-party files under 1500-line maintainability limit
controls.py (1567) and test_workspaces_observability_part_001.py (1511) both
exceeded MAX_FIRST_PARTY_FILE_LINES, failing
test_first_party_code_files_stay_under_line_limit. Extract the service-driven
stack-release helpers (_release_runtime_stack, _finalize_service_stack_release,
_stack_stop_error_from_cleanup) into controls_stack_release.py wired via
_WorkspaceStackReleaseMixin (mirroring _WorkspaceGuideMixin), and split the
observability test module's tail into part_004. Behavior unchanged.
